requirements.txt
openai
python-dotenv
tqdm
pydantic
google-genai